fix for Shapely 2.0
