+++ requirements.txt
@@ -4,7 +4,6 @@ pandas>=1.4
 astropy>=5.0
 astroquery>=0.4
 matplotlib>=3.5
-shapely>=1.6
-descartes>=1.1
+shapely>=2.0
 jax>=0.3
 jaxlib>=0.3
