+++ warpfield/telescope/util.py
@@ -4,11 +4,38 @@
 '''
 
 from astropy.wcs import WCS
+import numpy as np
 
 __debug_mode__ = False
+__arcsec_to_um__ = 1.0 / 3600
 
 
-def get_projection(pointing, scale=1.0 / 3600., lhcs=True, name='TAN'):
+def estimate_frame_from_ctype(ctype):
+    ''' Estimate a coordinate frame from CTYPE.
+
+    Arguments:
+      ctype (tuple):
+          A tuple of CTYPE strings.
+
+    Returns:
+      A string to specify the coordinate frame.
+    '''
+    ctype1, ctype2 = ctype
+
+    if 'GLON' in ctype1 and 'GLAT' in ctype2:
+        return 'galactic'
+    elif 'RA' in ctype1 and 'DEC' in ctype2:
+        return 'icrs'
+    else:
+        raise ValueError(f'unsupported CTYPE: {ctype}')
+
+
+def get_projection(
+        pointing,
+        scale=__arcsec_to_um__,
+        rotation=0.0,
+        left_hand_system=True,
+        projection='TAN'):
     ''' Obtain the gnomonic projection instance.
 
     Arguments:
@@ -36,16 +63,17 @@ def get_projection(pointing, scale=1.0 / 3600., lhcs=True, name='TAN'):
     if pointing.frame.name == 'galactic':
         lon = pointing.galactic.l.deg
         lat = pointing.galactic.b.deg
-        proj.wcs.ctype = [f'GLON-{name}', f'GLAT-{name}']
+        proj.wcs.ctype = [f'GLON-{projection}', f'GLAT-{projection}']
     else:
         lon = pointing.icrs.ra.deg
         lat = pointing.icrs.dec.deg
-        proj.wcs.ctype = [f'RA---{name}', f'DEC--{name}']
+        proj.wcs.ctype = [f'RA---{projection}', f'DEC--{projection}']
     proj.wcs.crval = [lon, lat]
-    if lhcs is True:
-        proj.wcs.cdelt = [-scale, scale]
-    else:
-        proj.wcs.cdelt = [scale, scale]
+    rot = np.array([
+        [+np.cos(rotation), +np.sin(rotation)],
+        [-np.sin(rotation), +np.cos(rotation)]])
+    delt = np.diag([-scale, scale] if left_hand_system else [scale, scale])
+    proj.wcs.cd = rot @ delt
     proj.wcs.cunit = ['deg', 'deg']
 
     if __debug_mode__ is True:
